add explicit test for #344
move tests for #222 #322 #327 #328 #35 #37 #38 #39 to dedicated package
