+++ src/test/java/net/openhft/chronicle/wire/issue/WireBug38Test.java
@@ -1,7 +1,8 @@
-package net.openhft.chronicle.wire;
+package net.openhft.chronicle.wire.issue;
 
 import net.openhft.chronicle.bytes.Bytes;
 import net.openhft.chronicle.core.io.IORuntimeException;
+import net.openhft.chronicle.wire.*;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.junit.Test;
@@ -11,31 +12,26 @@ import java.nio.ByteBuffer;
 import static junit.framework.TestCase.assertEquals;
 import static org.junit.Assert.assertNotNull;
 
-public class WireBug37Test extends WireTestCommon {
+public class WireBug38Test extends WireTestCommon {
     @Test
-    public void testNewlineInString() {
+    public void testNestedObj() {
         @NotNull final WireType wireType = WireType.TEXT;
-        @NotNull final String exampleString = "hello\nworld";
+        @NotNull final String exampleString = "{";
 
-        @NotNull final MarshallableObj obj1 = new MarshallableObj();
-        @NotNull final MarshallableObj obj2 = new MarshallableObj();
-        @NotNull final MarshallableObj obj3 = new MarshallableObj();
+        @NotNull final Outer obj1 = new Outer();
+        @NotNull final Outer obj2 = new Outer();
 
-        obj1.append(exampleString);
-        obj2.append(exampleString);
-
-        assertEquals(obj1, obj2);
+        obj1.getObj().append(exampleString);
 
         final Bytes<ByteBuffer> bytes = Bytes.elasticByteBuffer();
-        obj2.writeMarshallable(wireType.apply(bytes));
+        obj1.writeMarshallable(wireType.apply(bytes));
 
         final String output = bytes.toString();
        // System.out.println("output: [" + output + "]");
 
-        obj3.readMarshallable(wireType.apply(Bytes.from(output)));
+        obj2.readMarshallable(wireType.apply(Bytes.from(output)));
 
-        assertEquals(obj2, obj3);
-
+        assertEquals(obj1, obj2);
         bytes.releaseLast();
     }
 
@@ -80,6 +76,40 @@ public class WireBug37Test extends WireTestCommon {
         @Override
         public String toString() {
             return builder.toString();
+        }
+    }
+
+    static class Outer implements Marshallable {
+        private final MarshallableObj obj = new MarshallableObj();
+
+        @NotNull
+        MarshallableObj getObj() {
+            return obj;
+        }
+
+        @Override
+        public void readMarshallable(@NotNull WireIn wire) throws IORuntimeException {
+            wire.read(() -> "obj").marshallable(obj);
+        }
+
+        @Override
+        public void writeMarshallable(@NotNull WireOut wire) {
+            wire.write(() -> "obj").marshallable(obj);
+        }
+
+        @Override
+        public boolean equals(@Nullable Object o) {
+            if (this == o) return true;
+            if (o == null || getClass() != o.getClass()) return false;
+
+            @Nullable Outer outer = (Outer) o;
+
+            return obj.equals(outer.obj);
+        }
+
+        @Override
+        public int hashCode() {
+            return obj.hashCode();
         }
     }
 }
